+++ Changelog.md
@@ -2,6 +2,7 @@
 
 - Add --delete parameter
 - Change progress message
+- Fix --version parameter
 
 
 # v0.0.1